size_t and int repair

+++ include/libft.h
@@ -6,7 +6,7 @@
 /*   By: gcoqueir <gcoqueir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/12 08:39:51 by gcoqueir          #+#    #+#             */
-/*   Updated: 2023/06/14 13:47:22 by gcoqueir         ###   ########.fr       */
+/*   Updated: 2023/06/20 14:13:32 by gcoqueir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 # define UPPER_HEX "0123456789ABCDEF"
 # define LOWER_HEX "0123456789abcdef"
 
-size_t	ft_strlen(char *s);
+int		ft_strlen(char *s);
 void	*ft_memset(void *s, int c, size_t n);
 void	ft_bzero(void *s, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
